Refactor: Encapsulate CCA adding logic in Person class
Introduced Person.addCca(CcaInformation) to handle the creation
of a new Person instance with an added CCA. Updated AddCcaToStudentCommand
to use this method instead of manually constructing the new Person.

This improves encapsulation by moving the responsibility of creating
the modified Person state into the Person class itself, reducing
coupling with the command.